ToWebP/requirements.txt
Image==1.5.33
Pillow==8.4.0
